+++ R/methods_Rchoice.R
@@ -1,0 +1,10 @@
+#' @rdname set_coef
+#' @export
+set_coef.hetprob <- function(model, coefs, ...) {
+    model[["estimate"]][names(coefs)] <- coefs
+    model
+}
+
+#' @rdname set_coef
+#' @export
+set_coef.ivpml <- set_coef.hetprob